jobs-hunt-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "job-hunt-agent"
version = "0.1.0"
description = "自动化求职 AI Agent：抓取岗位 → 智能筛选 → 定制简历 → 自动投递"
requires-python = ">=3.11"

dependencies = [
    # LLM 工作流编排
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",

    # Anthropic SDK
    "anthropic>=0.40.0",

    # 网页自动化
    "playwright>=1.49.0",

    # 数据模型与配置验证
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",

    # 数据库
    "aiosqlite>=0.20.0",

    # 简历处理
    "python-docx>=1.1.0",
    "weasyprint>=62.0",
    "markdown>=3.7",

    # 工具库
    "tenacity>=9.0.0",      # 重试逻辑
    "rich>=13.0.0",         # 控制台美化输出
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
]

[project.scripts]
job-hunt = "main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
